fix: reduce code duplication
